split into multiple files and move typings into its own file
