+++ services/generic-proxy/.eslintrc.yml
@@ -1,0 +1,4 @@
+rules:
+    arrow-parens:
+        - error
+        - always
